+++ engine/src/main/java/net/jqwik/engine/properties/shrinking/ShrinkableBigDecimal.java
@@ -14,8 +14,16 @@ public class ShrinkableBigDecimal extends AbstractShrinkable<BigDecimal> {
 	private final BigDecimal target;
 	private final BigDecimalShrinkingCandidates shrinkingCandidates;
 
-	public ShrinkableBigDecimal(BigDecimal value, Range<BigDecimal> range, int scale) {
-		this(value, range, scale, determineTarget(range, value));
+	public static BigDecimal defaultShrinkingTarget(BigDecimal value, Range<BigDecimal> range) {
+		if (range.includes(BigDecimal.ZERO))
+			return BigDecimal.ZERO;
+		else {
+			if (value.compareTo(BigDecimal.ZERO) < 0)
+				return range.max;
+			if (value.compareTo(BigDecimal.ZERO) > 0)
+				return range.min;
+		}
+		return value; // Should never get here
 	}
 
 	public ShrinkableBigDecimal(BigDecimal value, Range<BigDecimal> range, int scale, BigDecimal shrinkingTarget) {
@@ -23,15 +31,17 @@ public class ShrinkableBigDecimal extends AbstractShrinkable<BigDecimal> {
 		this.range = range;
 		this.scale = scale;
 		this.target = shrinkingTarget;
+		checkTargetInRange(shrinkingTarget);
+		checkValueInRange(value);
 		this.shrinkingCandidates = new BigDecimalShrinkingCandidates(this.range, this.target);
 	}
 
 	@Override
 	public Set<Shrinkable<BigDecimal>> shrinkCandidatesFor(Shrinkable<BigDecimal> shrinkable) {
 		return shrinkingCandidates.candidatesFor(shrinkable.value())
-			.stream() //
-			.map(aBigDecimal -> new ShrinkableBigDecimal(aBigDecimal, range, scale)) //
-			.collect(Collectors.toSet());
+								  .stream() //
+								  .map(aBigDecimal -> new ShrinkableBigDecimal(aBigDecimal, range, scale, defaultShrinkingTarget(aBigDecimal, range))) //
+								  .collect(Collectors.toSet());
 	}
 
 	@Override
@@ -47,20 +57,18 @@ public class ShrinkableBigDecimal extends AbstractShrinkable<BigDecimal> {
 		return bigIntegerDistance.append(decimalDistance);
 	}
 
-	private static BigDecimal determineTarget(Range<BigDecimal> range, BigDecimal value) {
+	private void checkTargetInRange(BigDecimal value) {
+		if (!range.includes(value)) {
+			String message = String.format("Shrinking target <%s> is outside allowed range %s", value, range);
+			throw new JqwikException(message);
+		}
+	}
+
+	private void checkValueInRange(BigDecimal value) {
 		if (!range.includes(value)) {
 			String message = String.format("Number <%s> is outside allowed range %s", value, range);
 			throw new JqwikException(message);
 		}
-		if (range.includes(BigDecimal.ZERO))
-			return BigDecimal.ZERO;
-		else {
-			if (value.compareTo(BigDecimal.ZERO) < 0)
-				return range.max;
-			if (value.compareTo(BigDecimal.ZERO) > 0)
-				return range.min;
-		}
-		return value; // Should never get here
 	}
 
 }