+++ docs/release-notes.md
@@ -9,6 +9,7 @@ title: jqwik Release Notes
 <!-- DON'T EDIT THIS SECTION, INSTEAD RE-RUN doctoc TO UPDATE -->
 ### Table of Contents  
 
+- [1.1.5](#115)
 - [1.1.4](#114)
 - [1.1.3](#113)
 - [1.1.2](#112)
@@ -19,6 +20,17 @@ title: jqwik Release Notes
 - [0.8.x](#08x)
 
 <!-- END doctoc generated TOC please keep comment here to allow auto update -->
+
+## 1.1.5
+
+<p style="padding-left:1em;font-size:larger">
+<a href="/docs/1.1.5/user-guide.html">User guide</a>
+and
+<a href="/docs/1.1.5/javadoc/index.html">javadoc</a>
+</p>
+
+- Added `shrinkTowards(value)` to all decimal number arbitraries: 
+  FloatArbitrary, DoubleArbitrary, BigDecimalArbitrary
 
 ## 1.1.4